requirements.txt
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
paramiko==2.7.1
pycparser==2.19
PyNaCl==1.3.0
pysftp==0.2.9
PyYAML==5.3
six==1.14.0
